Created and tested the superPCA_pVals() function as a wrapper function for almost the entire supervised PCA workflow. We tested it on the 177 tumour set and the 250 colon set, but haven't tested it on the 58 ovarian set.
Edited the expressedOmes() function to have the possibility of returning an Omics* object. This was it can be used in a workflow to clean pathways or it can be called from within the extract_aesPCs() function.

Also, we fixed the trim and min.features arguments to be the same (3), and we found that our version of the Cox model (coxTrain_fun()) can't handle missing values.