Added the next few James comments to the README.
Added functionality to return the trimmed set size as a vector of the same length as the original set size (even if pathways are removed). Updated the comments in AES-PCA vignette to explain that the setsize pathway list element is added at object creation. Updated the matrix class checks (== doesn't work for objects with multiple classes).