Many of the internal functions require the pathway list to be named. I have corrected this issue in the Omics* object creation step.
The pathway name vector will be checked for existence, completeness, and uniqueness; each issue is resolved. The clean names vector is then attached to the TERMS and setsize vectors and reattached to the pathway list.

Also, I updated the help documentation to reflect the code edits of the last committ (to the creation suite and expressedOmes() functions).